Use the Command design pattern for pressing keys.
TWRTypingCommand is the abstract superclass that provides the functionality of pressing keys to its subclasses. Subclasses override -key to provide which key should be pressed.

TWREnterCharacterCommand locates the correct key to press for the given character, which can be pressed by calling -execute.
TWRPressSpaceCommand enters a space character when -execute is called.
TWRPressShiftCommand presses the shift key when -execute is called, switching the case being used by the keyboard.

Moved the -subviewsPassingTest: UIView convenience method into UIApplication+TWRKeyboard as it is the only place the method is used, keeping things clean.